+++ HISTORY.md
@@ -25,6 +25,7 @@
         * new "tts_started" extension, reports (in a separate process) whether the user has started speaking in response to a "tts_recognize" query
         * TTS library: new "started speech response?" predicate
     * always "normalize" SVGs on import, avoids "cut-off" costume parts
+    * new general "dta_export(data, name, type)" extension (for csv, json, etc.)
 * **Notable Fixes:**
     * frequency distribution analysis library: fixed "plot bars" block to handle zero values gracefully
 * **Translation Updates:**
@@ -38,6 +39,7 @@
 * fancy-text-morphs: added "jshint: esversion: 11" comment (for use with JSHint)
 * fancy-text-morphs: added "global" comment declcaration (for use with JSHint)
 * fancy-text-morphs: fixed some variable declaration / usage glitches
+* extensions: added general "dta_export(data, name, type)" extension (for csv, json, etc.)
 
 ### 2025-12-13
 * frequency distribution analysis library: fixed "plot bars" block to handle zero values gracefully